Disable failing test on azure (#319)
