+++ test/integration/recover_nse_heal_test.go
@@ -1,4 +1,4 @@
-// +build recover, !gke
+// +build recover,!gke
 
 package nsmd_integration_tests
